requirements.txt
numpy                     1.18.5                   
pandas                    1.2.0              
pytorch-fast-transformers 0.4.0                     
scikit-learn              0.24.0                          
torch                     1.6.0              
torch-cluster             1.5.8              
torch-geometric           1.6.3              
torch-scatter             2.0.5              
torch-sparse              0.6.8              
torch-spline-conv         1.2.0                        